chaged a variable type so that it would work better with raylib without needing to cast the type

+++ collision_objects.h
@@ -22,7 +22,7 @@ typedef struct {
 typedef struct {
 	Object objects[MAX_OBJECTS];
 	int obj_num;
-	float x[MinMax],y[MinMax];
+	int x[MinMax],y[MinMax];
 
 } Environment;
